generic absolute import rule
