+++ lib/rules/common-absolute-import.ts
@@ -2,57 +2,9 @@
  * @fileoverview Ensure that imports from common/ are absolute instead of relative.
  */
 
-import { ESLintUtils } from "@typescript-eslint/experimental-utils";
+import absoluteImportRule from "../absolute-import-rule";
 
-const rule = ESLintUtils.RuleCreator(
-  (ruleName) =>
-    `https://github.com/VantaInc/eslint-plugin-vanta/blob/main/docs/rules/${ruleName}.md`
-)<unknown[], "default">({
-  name: "common-absolute-import",
-  meta: {
-    fixable: "code",
-    docs: {
-      description: "Import common/ from its absolute path",
-      recommended: "error",
-    },
-    messages: {
-      default: "Import code in common/ from its absolute path",
-    },
-    type: "suggestion",
-    schema: [],
-  },
-  defaultOptions: [],
-  create(context) {
-    const badCommonImportRegex = /\.\.\/.*\/?(common\/)src\/(.*)/s;
-    return {
-      ImportDeclaration(node) {
-        if (typeof node.source.value !== "string") {
-          return;
-        }
-        const matches = node.source.value.match(badCommonImportRegex);
-        if (!matches) {
-          return;
-        }
-        const replacement = node.source.value.replace(
-          badCommonImportRegex,
-          "$1$2"
-        );
-        context.report({
-          node,
-          messageId: "default",
-          fix(fixer) {
-            return [
-              fixer.replaceTextRange(
-                [node.source.range[0] + 1, node.source.range[1] - 1],
-                replacement
-              ),
-            ];
-          },
-        });
-      },
-    };
-  },
-});
+const rule = absoluteImportRule("common");
 
 module.exports = rule;
