+++ lib/rules/server-common-absolute-import.ts
@@ -2,59 +2,10 @@
  * @fileoverview Ensure that imports from server-common/ are absolute instead of relative.
  */
 
- import { ESLintUtils } from "@typescript-eslint/experimental-utils";
+import absoluteImportRule from "../absolute-import-rule";
 
- const rule = ESLintUtils.RuleCreator(
-   (ruleName) =>
-     `https://github.com/VantaInc/eslint-plugin-vanta/blob/main/docs/rules/${ruleName}.md`
- )<unknown[], "default">({
-   name: "server-common-absolute-import",
-   meta: {
-     fixable: "code",
-     docs: {
-       description: "Import server-common/ from its absolute path",
-       recommended: "error",
-     },
-     messages: {
-       default: "Import code in server-common/ from its absolute path",
-     },
-     type: "suggestion",
-     schema: [],
-   },
-   defaultOptions: [],
-   create(context) {
-     const badCommonImportRegex = /\.\.\/.*\/?(server-common\/)src\/(.*)/s;
-     return {
-       ImportDeclaration(node) {
-         if (typeof node.source.value !== "string") {
-           return;
-         }
-         const matches = node.source.value.match(badCommonImportRegex);
-         if (!matches) {
-           return;
-         }
-         const replacement = node.source.value.replace(
-           badCommonImportRegex,
-           "$1$2"
-         );
-         context.report({
-           node,
-           messageId: "default",
-           fix(fixer) {
-             return [
-               fixer.replaceTextRange(
-                 [node.source.range[0] + 1, node.source.range[1] - 1],
-                 replacement
-               ),
-             ];
-           },
-         });
-       },
-     };
-   },
- });
- 
- module.exports = rule;
- 
- export default rule;
- 
+const rule = absoluteImportRule("server-common");
+
+module.exports = rule;
+
+export default rule;
